cart added commit from Elmurodov

+++ src/main/java/sarik/dev/foodwaveproject/entity/CartItem.java
@@ -25,7 +25,7 @@ public class CartItem {
 	private Product product;
 
 	private Integer quantity;
-	private double discount;
-	private double productPrice;
+	private long discount;
+	private long productPrice;
 
 }
